+++ source/client/application/settings/settings.card.js
@@ -1,10 +1,10 @@
 import React from "react"
 import { Avatar, Card, CardHeader, CardText, FontIcon } from "material-ui"
 import { lightBlue500 } from "material-ui/styles/colors"
-import { shieldIconStyle } from "./advanced.styles"
-import { AdvancedFormItemsComponent } from "./advanced.form.items.component"
+import { shieldIconStyle } from "./settings.styles"
+import { SettingsCardItems } from "./settings.card.items"
 
-export class AdvancedFormComponent extends React.Component {
+export class SettingsCard extends React.Component {
     render() {
         const avatar = <Avatar backgroundColor={ lightBlue500 } icon={ <FontIcon style={ shieldIconStyle } className="fa fa-shield"/> }/>
 
@@ -12,7 +12,7 @@ export class AdvancedFormComponent extends React.Component {
             <Card>
                 <CardHeader title="Password Settings" subtitle="Customize your generated password" avatar={ avatar }/>
                 <CardText>
-                    <AdvancedFormItemsComponent/>
+                    <SettingsCardItems/>
                 </CardText>
             </Card>
         )
